refactor(cart): Cart 페이지의 레이아웃(main, section)과 스타일을 pages 레이어로 이동하여 책임 분리
