+++ src/app/(house)/cart/page.tsx
@@ -2,16 +2,9 @@
 
 import { withAuth } from "@/shared/hocs/withAuth";
 import { CartPage } from "@/pages/cart/ui/CartPage";
-import * as s from "@/app/(house)/cart/style.css";
 
 function Page() {
-  return (
-    <main>
-      <section className={s.cartSection}>
-        <CartPage />
-      </section>
-    </main>
-  );
+  return <CartPage />;
 }
 
 export default withAuth(Page, "protected");
